Mysql implemented

+++ demo/src/main/java/com/wipro/raemisclient/microservicetemplate/SubscriberDAO.java
@@ -1,15 +1,22 @@
-package com.wipro.raemisclient.dao;
+package com.wipro.raemisclient.microservicetemplate;
 
-import com.wipro.raemisclient.common.Core5GDetails;
-import com.wipro.raemisclient.model.Subscriber;
-import java.sql.*;
+import java.io.IOException;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-public class SubscriberDAO implements DAOInterface<Subscriber> {
+import com.wipro.raemisclient.apiservice.NotifyMessage;
+import com.wipro.raemisclient.common.Core5GDetails;
+import com.wipro.raemisclient.model.response.SubscriberResponse;
+
+public class SubscriberDAO implements DAOInterface<SubscriberResponse> {
 
 	private static final String DELETE_SUBSCRIBER_QUERY = "DELETE FROM subscriber where imsi=";
 	private static final String INSERT_SUBSCRIBER_QUERY = "INSERT INTO subscriber VALUES";
@@ -22,14 +29,14 @@ public class SubscriberDAO implements DAOInterface<Subscriber> {
 	}
 
 	@Override
-	public void insertRecords(List<Subscriber> listOfData) throws SQLException {
-		for (Subscriber data : listOfData) {
+	public void insertRecords(List<SubscriberResponse> listOfData) throws SQLException {
+		for (SubscriberResponse data : listOfData) {
 			insertRecord(data);
 		}
 	}
 
 	@Override
-	public void insertRecord(Subscriber data) throws SQLException {
+	public void insertRecord(SubscriberResponse data) throws SQLException {
 		try (Statement statement = connection.createStatement()) {
 			String queryParam = "(" + data.getId() + ", '" + data.getImsi() + "', '" + data.getTmsi() + "', '"
 					+ data.getPtmsi() + "'," + " '" + data.getImei() + "', '" + data.getMsisdn() + "', '"
@@ -49,7 +56,7 @@ public class SubscriberDAO implements DAOInterface<Subscriber> {
 	}
 
 	@Override
-	public void updateRecord(Subscriber data) throws SQLException {
+	public void updateRecord(SubscriberResponse data) throws SQLException {
 		try {
 			PreparedStatement preparedStmt = connection.prepareStatement(UPDATE_SUBSCRIBER_QUERY);
 			preparedStmt.setString(1, data.getTmsi());
@@ -73,8 +80,8 @@ public class SubscriberDAO implements DAOInterface<Subscriber> {
 	}
 
 	@Override
-	public List<Subscriber> getRecordByParam(Map<String, Object> paramMap) throws SQLException {
-		List<Subscriber> subscribers = new ArrayList<Subscriber>();
+	public List<SubscriberResponse> getRecordByParam(Map<String, Object> paramMap) throws SQLException {
+		List<SubscriberResponse> subscribers = new ArrayList<SubscriberResponse>();
 		try (Statement statement = connection.createStatement()) {
 			StringBuffer sb = new StringBuffer(" where ");
 			int mapSize = paramMap.size();
@@ -93,7 +100,7 @@ public class SubscriberDAO implements DAOInterface<Subscriber> {
 			ResultSet resultSet = statement.executeQuery(VIEW_SUBSCRIBER_QUERY + sb);
 
 			while (resultSet.next()) {
-				Subscriber subscriber = new Subscriber();
+				SubscriberResponse subscriber = new SubscriberResponse();
 				subscriber.setImsi(resultSet.getString(2));
 				subscriber.setImei(resultSet.getString(5));
 				subscriber.setLocal_ps_attachment(resultSet.getString(9));
@@ -107,15 +114,15 @@ public class SubscriberDAO implements DAOInterface<Subscriber> {
 	}
 
 	@Override
-	public void updateOrInsertRecords(List<Subscriber> listOfData) throws SQLException {
+	public void updateOrInsertRecords(List<SubscriberResponse> listOfData) throws SQLException, IOException {
 
 		if (listOfData == null || listOfData.isEmpty())
 			return;
 
 		Map<String, Object> paramMap = new HashMap<String, Object>();
 		paramMap.put("core_id", Core5GDetails._5G_CORE_ID);
-		List<Subscriber> existingSubscribers = getRecordByParam(paramMap);
-
+		List<SubscriberResponse> existingSubscribers = getRecordByParam(paramMap);
+		boolean isDataUpdated = false;
 		if (existingSubscribers.isEmpty()) {
 			insertRecords(listOfData);
 		} else {
@@ -127,19 +134,32 @@ public class SubscriberDAO implements DAOInterface<Subscriber> {
 
 			List<String> deleteSubscribers = existingSubscribersImsi.stream()
 					.filter(i -> !currentSubscribersImsi.contains(i)).collect(Collectors.toList());
-			deleteRecords(deleteSubscribers);
-			List<Subscriber> insertSubscribers = listOfData.stream()
-					.filter(i -> !existingSubscribersImsi.contains(i.getImsi())).collect(Collectors.toList());
-			insertRecords(insertSubscribers);
 
-			for (Subscriber curr_sub : listOfData) {
-				for (Subscriber ext_sub : existingSubscribers) {
+			if (deleteSubscribers != null && !deleteSubscribers.isEmpty()) {
+				deleteRecords(deleteSubscribers);
+				isDataUpdated = true;
+			}
+			List<SubscriberResponse> insertSubscribers = listOfData.stream()
+					.filter(i -> !existingSubscribersImsi.contains(i.getImsi())).collect(Collectors.toList());
+
+			if (insertSubscribers != null && !insertSubscribers.isEmpty()) {
+				insertRecords(insertSubscribers);
+				isDataUpdated = true;
+			}
+
+			for (SubscriberResponse curr_sub : listOfData) {
+				for (SubscriberResponse ext_sub : existingSubscribers) {
 					if (curr_sub.getImsi().equals(ext_sub.getImsi())) {
-						if (!curr_sub.getLocal_ps_attachment().equals(ext_sub.getLocal_ps_attachment()))
+						if (!curr_sub.getLocal_ps_attachment().equals(ext_sub.getLocal_ps_attachment())) {
 							updateRecord(curr_sub);
+							isDataUpdated = true;
+						}
 					}
 				}
 			}
+		}
+		if (isDataUpdated) {
+			NotifyMessage.sendMessage_Subscriber();
 		}
 	}
 
@@ -155,7 +175,7 @@ public class SubscriberDAO implements DAOInterface<Subscriber> {
 	}
 
 	@Override
-	public void pollRecords(List<Subscriber> listOfData) throws SQLException, InterruptedException {
+	public void pollRecords(List<SubscriberResponse> listOfData) throws SQLException, InterruptedException, IOException {
 		updateOrInsertRecords(listOfData);
 		System.out.println("Subscriber records are polling..");
 	}
